Move LTS-related test helpers to separate file
Move some things out of the giant TimeStepperTestUtils.
