+++ tests/Unit/Helpers/Time/TimeSteppers/CMakeLists.txt
@@ -4,6 +4,7 @@
 set(LIBRARY "TimeStepperHelpers")
 
 set(LIBRARY_SOURCES
+  LtsHelpers.cpp
   RungeKutta.cpp
   TimeStepperTestUtils.cpp
   )
